requirements.txt
torch
numpy
pillow